Refactored POMs towards consistency
All duplicate information was moved to the parent pom, the description of the parent POM
is no longer the same as the Wicket CDI module.

All dependency information is now  centralized in the parent POM, including the plugins
necessary for building.

I've included the enforcer plugin for consistent POM layout (which is annoying, but helps in the
end for enforcing a single style across all project modules).

All plugins now use their most recent version.